주문 조회, 회원의 주문내역들 조회, 회원 주문 구현

+++ src/main/java/shop/sajotuna/order/orders/domain/OrderPackaging.java
@@ -1,8 +1,6 @@
 package shop.sajotuna.order.orders.domain;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,7 +14,7 @@ import lombok.Setter;
 public class OrderPackaging {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "packaging", nullable = false)
     private String packaging;
